reformat(move): moved some dtos to auth-dtos folder
